requirements.txt
# Django Core
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9

# Async Tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# Image Processing
Pillow==10.1.0

# Environment & Security
python-dotenv==1.0.0
django-cors-headers==4.3.1

# Utilities
python-decouple==3.8
requests==2.31.0
